Agent/python-face-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.0
opencv-python==4.10.0.84
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
# InsightFace and onnxruntime are optional - install separately if needed
# For full face recognition, uncomment and install:
# insightface==0.7.3
# onnxruntime>=1.16.0
# 
# To install: pip install insightface onnxruntime

